buy-wise-microservice/requirements.txt
Flask==2.3.3
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
gunicorn==21.2.0